refactor: move types to new types dir with non-relative module imports
