--- src/pages/game/events/spawn-event.ts
@@ -1,8 +1,0 @@
-import type { Component } from '../simulation/ecs/component.interface';
-import type { AbstractEvent } from './abstract-event';
-
-export interface SpawnEvent extends AbstractEvent {
-    event: 'spawn';
-    ts: string;
-    components: Component[];
-}
