--- src/pages/game/simulation/ecs/component.interface.ts
@@ -1,5 +1,0 @@
-import type { ComponentType } from './component-type.enum';
-
-export interface Component {
-    type: ComponentType;
-}
