+++ src/pages/game/simulation/ecs/ecs-registry.ts
@@ -1,4 +1,4 @@
-import type { Component } from './component.interface';
+import type { Component } from 'types/global/component';
 
 export interface Entity {
     id: number;
