+++ src/pages/game/simulation/ecs/renderer.component.ts
@@ -1,4 +1,4 @@
-import { ComponentType } from './component-type.enum';
+import { ComponentType } from 'types/global/componentType.enum';
 
 export interface UniformOverrides {
     [name: string]: unknown;
