+++ src/pages/game/simulation/ecs/transform.component.ts
@@ -1,7 +1,7 @@
-import { ComponentType } from './component-type.enum';
+import { ComponentType } from 'types/global/componentType.enum';
 
 import type { vec3 } from 'gl-matrix';
-import type { Component } from './component.interface';
+import type { Component } from 'types/global/component';
 
 export class TransformComponent implements Component {
     public type: ComponentType.TRANSFORM = ComponentType.TRANSFORM;
