+++ src/pages/game/simulation/simulation-internals.interface.ts
@@ -1,5 +1,5 @@
 import type EventEmitter from 'events';
-import type { AbstractEvent } from '../events/abstract-event';
+import type { AbstractEvent } from 'types/events';
 import type { CanvasIO } from './canvas-io';
 import type { EcsRegistry } from './ecs/ecs-registry';
 import type { SceneRenderer } from './renderer/scene-renderer';
