+++ src/pages/game/simulation/simulation.interface.ts
@@ -1,5 +1,5 @@
 import type { vec4 } from 'gl-matrix';
-import type { AbstractEvent } from '../events/abstract-event';
+import type { AbstractEvent } from 'types/events';
 
 export interface Simulation {
     run(canvasEl: HTMLCanvasElement, bgColor: vec4): Promise<void>;
