+++ types/events/abstractEvent.ts
