+++ types/events/index.ts
@@ -1,0 +1,5 @@
+export type { AbstractEvent } from './abstractEvent';
+export type { ChatEvent } from './chatEvent';
+export type { MoveEvent } from './moveEvent';
+export type { SpawnEvent } from './spawnEvent';
+export type { DespawnEvent } from './despawnEvent';
