+++ types/events/spawnEvent.ts
@@ -1,0 +1,8 @@
+import type { Component } from 'types/global/component';
+import type { AbstractEvent } from './abstractEvent';
+
+export interface SpawnEvent extends AbstractEvent {
+    event: 'spawn';
+    ts: string;
+    components: Component[];
+}
