+++ types/global/component.ts
@@ -1,0 +1,5 @@
+import type { ComponentType } from './componentType.enum';
+
+export interface Component {
+    type: ComponentType;
+}
