+++ types/global/componentType.enum.ts
